app/requirements.txt
fastapi
uvicorn
google-auth
google-api-python-client
python-dotenv
